continuing with putting qt code in functions
